Data-Pipeline/requirements.txt
# =============================================================================
# Wildfire MLOps Pipeline — Python Dependencies
# =============================================================================
# Unpinned here; use with constraints.txt for reproducible installs:
#   pip install -r requirements.txt -c constraints.txt
#
# Install:
#   pip install -r requirements.txt -c constraints.txt
#
# Why constraints.txt instead of pinning here:
#   Airflow 2.8.1 ships its own constraint file per Python version.
#   Hard-pinning conflicts with it on macOS Arm64 (M-series) and some
#   Windows/conda environments. The constraint file bounds the dangerous
#   packages (pyarrow, numpy) while letting the resolver handle the rest.
#   pyarrow IS hard-pinned in constraints.txt because Parquet schema
#   encoding changed between major versions — all team members must match.
# =============================================================================

# --- Airflow ---
apache-airflow[postgres]==2.8.1

# --- Data stack ---
pandas
numpy
packaging
pyarrow

# --- HRRR weather (emergency/active watchdog triggers only) ---
herbie-data          # HRRR S3 path construction + cfgrib/xarray integration
cfgrib>=0.9.10       # eccodes-backed GRIB2 reading via xarray
eccodes              # system GRIB2 library (also needs apt: libeccodes-dev)
scipy                # RegularGridInterpolator for HRRR → H3 centroid extraction
xarray               # N-dimensional array support used by cfgrib/herbie-data


# --- Geo stack ---
h3
geopandas
shapely
rasterio
fiona
rasterstats

# --- HTTP / config ---
requests
pyyaml
python-json-logger

# --- Database ---
psycopg2-binary

# --- Data quality ---
great-expectations<1.0

# --- MLOps / Versioning ---
dvc
dvc-gs

# --- GCP ---
google-cloud-storage
google-auth

# --- ML Training ---
xgboost
lightgbm
scikit-learn
shap>=0.45,<0.50

# --- LLM Report Generation ---
google-genai
fastapi
uvicorn[standard]

# --- Testing ---
pytest
pytest-cov
pytest-mock

# --- Code quality ---
ruff
# hypothesis       # property-based testing